refacto: Using inputManager in GameController
