+++ Assets/Textures/pexels-felixmittermeier-956981.jpg.meta
@@ -147,7 +147,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      pexels-felixmittermeier-956981_0: -4943324347405272402
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
